backend/analyze-news/requirements.txt
psycopg2-binary>=2.9.0
groq>=0.4.0
